// src/types/generated/CoverageDeltas.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CoverageDeltas = { champion_rate_delta: number, matchup_delta: number, build_delta: number, discovered_delta: number, processed_delta: number, crawled_player_delta: number, elapsed_secs: number, };