// src/types/generated/CoverageRampSnapshotView.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type CoverageRampSnapshotView = { taken_at: number, champion_rate_rows: number, matchup_rows: number, build_rows: number, discovered_matches: number, fetched_matches: number, processed_matches: number, failed_matches: number, crawled_players: number, };