// src/types/generated/DataPipelineRefreshSummary.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type DataPipelineRefreshSummary = { before_status: string, after_status: string, actions: Array<string>, ddragon_champions: number, meraki_rates: number, builds_imported: number, matchups_imported: number, match_v5_matches: number, match_v5_rates: number, match_v5_matchups: number, match_v5_builds: number, match_v5_errors: number, data_pack_cached: boolean, cache_action: string, cache_promoted: boolean, errors: Array<string>, };