// src/types/generated/LiveCoverageRampReport.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CoverageRampReport } from "./CoverageRampReport";
import type { CoverageRampSnapshotView } from "./CoverageRampSnapshotView";
import type { DataPipelineRefreshSummary } from "./DataPipelineRefreshSummary";

export type LiveCoverageRampReport = { before: CoverageRampSnapshotView, after: CoverageRampSnapshotView, ramp: CoverageRampReport, refresh: DataPipelineRefreshSummary, champ_select_active: boolean, crawl_budget: number, };